+++ lib/src/internal/plugin.dart
@@ -11,6 +11,7 @@ import 'package:photo_manager/platform_utils.dart';
 
 import '../filter/base_filter.dart';
 import '../filter/path_filter.dart';
+import '../types/cancel_token.dart';
 import '../types/entity.dart';
 import '../types/thumbnail.dart';
 import '../types/types.dart';
@@ -22,14 +23,30 @@ import 'progress_handler.dart';
 PhotoManagerPlugin plugin = PhotoManagerPlugin();
 
 mixin BasePlugin {
-  MethodChannel _channel = const MethodChannel(PMConstants.channelPrefix);
+  MethodChannel _channel = const PMMethodChannel(PMConstants.channelPrefix);
 
   final Map<String, dynamic> onlyAddPermission = <String, dynamic>{
     'onlyAddPermission': true,
   };
 }
 
-class VerboseLogMethodChannel extends MethodChannel {
+class PMMethodChannel extends MethodChannel {
+  const PMMethodChannel(String name) : super(name);
+
+  @override
+  Future<T?> invokeMethod<T>(String method, [dynamic arguments]) {
+    if (arguments is! Map) {
+      return super.invokeMethod<T>(method, arguments);
+    }
+    arguments.putIfAbsent(
+      PMConstants.cancelTokenKey,
+      () => PMCancelToken().key,
+    );
+    return super.invokeMethod<T>(method, arguments);
+  }
+}
+
+class VerboseLogMethodChannel extends PMMethodChannel {
   VerboseLogMethodChannel({
     required String name,
     required this.logFilePath,
@@ -131,7 +148,7 @@ class PhotoManagerPlugin with BasePlugin, IosPlugin, AndroidPlugin, OhosPlugin {
         logFilePath: logPath,
       );
     } else {
-      _channel = const MethodChannel(PMConstants.channelPrefix);
+      _channel = const PMMethodChannel(PMConstants.channelPrefix);
     }
   }
 
@@ -241,17 +258,28 @@ class PhotoManagerPlugin with BasePlugin, IosPlugin, AndroidPlugin, OhosPlugin {
     }
   }
 
+  void _setCancelToken(
+    Map<String, dynamic> params,
+    PMCancelToken? cancelToken,
+  ) {
+    if (cancelToken != null) {
+      params[PMConstants.cancelTokenKey] = cancelToken.key;
+    }
+  }
+
   /// Get thumbnail of asset id.
   Future<typed_data.Uint8List?> getThumbnail({
     required String id,
     required ThumbnailOption option,
     PMProgressHandler? progressHandler,
+    PMCancelToken? cancelToken,
   }) {
     final Map<String, dynamic> params = <String, dynamic>{
       'id': id,
       'option': option.toMap(),
     };
     _injectProgressHandlerParams(params, progressHandler);
+    _setCancelToken(params, cancelToken);
     return _channel.invokeMethod(PMConstants.mGetThumb, params);
   }
 
@@ -259,12 +287,14 @@ class PhotoManagerPlugin with BasePlugin, IosPlugin, AndroidPlugin, OhosPlugin {
     String id, {
     PMProgressHandler? progressHandler,
     PMDarwinAVFileType? darwinFileType,
+    PMCancelToken? cancelToken,
   }) {
     final Map<String, dynamic> params = <String, dynamic>{
       'id': id,
       'darwinFileType': darwinFileType,
     };
     _injectProgressHandlerParams(params, progressHandler);
+    _setCancelToken(params, cancelToken);
     return _channel.invokeMethod(PMConstants.mGetOriginBytes, params);
   }
 
@@ -281,6 +311,7 @@ class PhotoManagerPlugin with BasePlugin, IosPlugin, AndroidPlugin, OhosPlugin {
     PMProgressHandler? progressHandler,
     int subtype = 0,
     PMDarwinAVFileType? darwinFileType,
+    PMCancelToken? cancelToken,
   }) async {
     final params = <String, dynamic>{
       'id': id,
@@ -289,6 +320,7 @@ class PhotoManagerPlugin with BasePlugin, IosPlugin, AndroidPlugin, OhosPlugin {
       'darwinFileType': darwinFileType?.value ?? 0,
     };
     _injectProgressHandlerParams(params, progressHandler);
+    _setCancelToken(params, cancelToken);
     return _channel.invokeMethod(PMConstants.mGetFullFile, params);
   }
 
@@ -488,6 +520,7 @@ class PhotoManagerPlugin with BasePlugin, IosPlugin, AndroidPlugin, OhosPlugin {
   Future<String?> getMediaUrl(
     AssetEntity entity, {
     PMProgressHandler? progressHandler,
+    PMCancelToken? cancelToken,
   }) async {
     if (PlatformUtils.isOhos) {
       return entity.id;
@@ -497,6 +530,7 @@ class PhotoManagerPlugin with BasePlugin, IosPlugin, AndroidPlugin, OhosPlugin {
       'type': entity.typeInt,
     };
     _injectProgressHandlerParams(params, progressHandler);
+    _setCancelToken(params, cancelToken);
     return _channel.invokeMethod(PMConstants.mGetMediaUrl, params);
   }
 
@@ -693,6 +727,18 @@ class PhotoManagerPlugin with BasePlugin, IosPlugin, AndroidPlugin, OhosPlugin {
       requestOption.toMap(),
     );
     return PermissionState.values[result];
+  }
+
+  Future<void> cancelRequest(PMCancelToken pmCancelToken) {
+    return _channel.invokeMethod(PMConstants.mCancelRequestWithCancelToken, {
+      'cancelToken': pmCancelToken.key,
+    });
+  }
+
+  Future<void> cancelAllRequest() {
+    return _channel.invokeMethod(
+      PMConstants.mCancelAllRequest,
+    );
   }
 }